Обновить испорты, экспорты для стор
