+++ src/components/FormConstructor/store/formElements/index.ts
@@ -1,6 +1,9 @@
-import { formConstructorSlice } from './slices'
-import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux'
-import { AppDispatch, RootState } from '..'
+import {
+  formConstructorSlice,
+  formConstructorReducer,
+  useAppDispatch,
+  useAppSelector,
+} from './slices'
 import {
   ButtonElementProps,
   LayoutElementProps,
@@ -8,11 +11,6 @@ import {
   FormElementUnion,
   GroupElementUnion,
 } from './types'
-
-const formConstructorReducer = formConstructorSlice.reducer
-
-const useAppDispatch = () => useDispatch<AppDispatch>()
-const useAppSelector: TypedUseSelectorHook<RootState> = useSelector
 
 export {
   formConstructorReducer,
